Web Animations: Fix bug with Document::getAnimations
Per spec, we should exclude effects that do not target an element in
the called document:

https://drafts.csswg.org/web-animations-1/#dom-document-getanimations

This CL also fixes an issue with the getAnimations.html test. It
previously presumed that GC would run between tests to collect animations
from previous tests, but that isn't required to happen. To avoid that
sort of issue, the tests were changed to explicitly finish their animations.

Bug: 828424
Change-Id: I41405d82184b17c1185931e34735a5f946573844